Disable AppCache for Active Backend Users
I had a customer that wanted to see the changes he made straight away. But with the enabled cache, this was not possible.
I talked to Toflar and we came to the solution that it would be a good idea to only enable the AppCache if the user is not logged in to the backend.