Fix "mail" DSN (see #1828)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | N/A
| Docs PR or issue | N/A

While working on a PR that replaces `symfony/swiftmailer-bundle` with `symfony/mailer` I noticed a small bug: the Contao Manager Plugin of the Manager Bundle currently checks for a `mailer_transport` parameter with the content `mail`, which is supposed to be changed to `sendmail` instead, since Swiftmailer dropped the support of the PHP `mail()` function.

However, this does not work in theory because it is done _after_ the DSN for the mailer was already built. So if you had `mailer_transport: mail` in your `parameters.yml`, it would use `smtp://…` instead of `sendmail://localhost`. It only works because during cache warming, `getExtensionConfig()` is called multiple times. So after the second time, `env(MAILER_URL)` will be correct again.

This PR adds the failing test for that and changes the order of execution to fix it.

Commits
-------

119f8189 fix "mail" DSN
f2ae165a Merge branch '4.9' into fix-mail-dsn
1c81c2fc Merge branch '4.9' into fix-mail-dsn