Do not set the server_version (see #3623)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #1696

For the record as described in https://github.com/contao/contao/pull/3623#issuecomment-952708988 the `CommandSchedulerListener` must only load the `Cron` service when it needs it. Otherwise the install tool would fail when terminating because the `Cron` service needs ORM metadata which itself needs to know the `server_version`.

Commits
-------

e071bfdc Do not set server_version to wrong value
2236de4a Skip the cache warmer only if the platform version is missing
1bda30f9 Merge branch '4.9' into fix/doctrine-server-version