Improve the system maintenance mode (see #3737)
Description
-----------

This is the complementary PR for https://github.com/contao/contao/pull/3729.

Once https://github.com/contao/contao/pull/3729 is merged, we'll have a better maintenance mode which can be enabled per root page and the content can be managed and styled individually.
This covers the "content maintenance mode" (aka system is running but parts of the website are just not ready yet).

What's missing in order to get rid of the lexik maintenance bundle completely is a new "system maintenance mode" (aka not even the system is running). This is a use case we have for example when updating a system, running migrations/new deployments etc.
But to make this a real improvement over the lexik maintenance `kernel.request` listener, we have to take this even a step further because it might be that the application isn't running at all because for example the whole `vendor` directory isn't even installed (yet). The maintenance message should still be shown then. 

That's why I have designed this new mode with a simple "if a given file exists, show and abort" approach directly in our `index.php` before even loading any vendors or autoloading files. That means you can also enable the "system maintenance mode" without having any CLI running. 

The implementation is based on a Twig template which defaults to `@ContaoCore/Error/service_unavailable.html.twig` so it's also backwards compatible ;-)

The new `contao:maintenance-mode` command takes `enable` or `disable` as arguments (can also use `on` or `off`).
When enabling, you can pass a custom Twig template name as well as custom template variables for maximum flexibility.

Commits
-------

4cccbaed Bye Lexik and maintenance mode
fd8db7bd Implemented a new system maintenance which doesn't require the system to work at all
03ae1443 CS
953e8911 Maintenance mode is now based on Twig
22d2ea9c Adjusted the command description
33217b92 CS
f6a0279f Revert the ServiceUnavailableException changes
2714b27a Write maintenance markup to var directory

Co-authored-by: Leo Feyer <github@contao.org>